+++ src/funcs/workflowsExecutionsGetWorkflowExecution.ts
@@ -4,7 +4,7 @@
  */
 
 import { MistralCore } from "../core.js";
-import { encodeSimple } from "../lib/encodings.js";
+import { encodeFormQuery, encodeSimple } from "../lib/encodings.js";
 import { matchStatusCode } from "../lib/http.js";
 import * as M from "../lib/matchers.js";
 import { compactMap } from "../lib/primitives.js";
@@ -103,6 +103,10 @@ async function $do(
     pathParams,
   );
 
+  const query = encodeFormQuery({
+    "include_search_keys": payload.include_search_keys,
+  });
+
   const headers = new Headers(compactMap({
     Accept: "application/json",
   }));
@@ -133,6 +137,7 @@ async function $do(
     baseURL: options?.serverURL,
     path: path,
     headers: headers,
+    query: query,
     body: body,
     userAgent: client._options.userAgent,
     timeoutMs: options?.timeoutMs || client._options.timeoutMs || 300000,